+++ app/src/main/java/nextstep/payments/designsystem/component/PaymentCard.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.designsystem.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
