+++ app/src/main/java/nextstep/payments/designsystem/theme/Color.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.ui.theme
+package nextstep.payments.designsystem.theme
 
 import androidx.compose.ui.graphics.Color
