+++ app/src/main/java/nextstep/payments/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.ui.theme
+package nextstep.payments.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
